requirements.txt
datasette
datasette-cluster-map
datasette-graphql
sqlite-utils
PyYAML
